+++ src/common/polyfills/all.js
@@ -1,0 +1,2 @@
+import './performance-now'
+
