--- src/plugins/plugins-config.json
@@ -1,24 +1,0 @@
-{
-  "youbora": {
-    "playerVersion": "{{pVersion}}",
-    "playerName": "{{pName}}",
-    "entryId": "{{entryId}}",
-    "entryName": "{{entryName}}",
-    "entryType": "{{entryType}}",
-    "sessionId": "{{sessionId}}",
-    "uiConfId": "{{uiConfId}}"
-  },
-  "kanalytics": {
-    "playerVersion": "{{pVersion}}",
-    "entryId": "{{entryId}}",
-    "entryType": "{{entryType}}",
-    "sessionId": "{{sessionId}}",
-    "ks": "{{ks}}",
-    "uiConfId": "{{uiConfId}}",
-    "partnerId": "{{partnerId}}"
-  },
-  "ima": {
-    "playerVersion": "{{pVersion}}",
-    "playerName": "{{pName}}"
-  }
-}
