--- src/polyfills/all.js
@@ -1,2 +1,0 @@
-import './performance-now'
-
